==> analysis/04-reduced-dimensions.Rmd <==
---
title: "Dimensionality reduction"
author: "James Ashmore"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

## Motivation

In this document we aim to reduce the number of separate dimensions in the data by performing dimenstionality reduction.

## Setup

Set chunk options:

```{r}
knitr::opts_chunk$set(
  autodep    = TRUE,
  cache      = TRUE,
  cache.path = "cache/04-reduced-dimensions.Rmd/",
  dev        = "png",
  error      = FALSE,
  message    = FALSE,
  warning    = FALSE
)
```

Load required packages:

```{r}
library(BiocSingular)
library(patchwork)
library(pheatmap)
library(scater)
library(scran)
```

Read experiment data:

```{r}
sce <- readRDS("data/03-feature-selection.rds")
```

## PCA

### Calculate PCA

Perform a principal components analysis (PCA) on the expression data:

```{r}
set.seed(409643924)

sce <- runPCA(sce, subset_row = rowData(sce)$selected, BSPARAM = ExactParam())
```

Cluster cells to help examine seperation of clusters by principal components:

```{r}
snn <- buildSNNGraph(sce, use.dimred = "PCA")

com <- igraph::cluster_walktrap(snn)

sce$cluster <- factor(com$membership)
```

### Find elbow

Find the elbow point in the percentage of variance explained by successive PCs:

```{r}
dim <- reducedDim(sce, "PCA")

var <- attr(dim, "percentVar")

num <- RclusTool::ElbowFinder(seq_along(var), var)
```

Plot the percentage of variance explained by successive PCs:

```{r}
dat <- data.frame(index = seq_along(var), total = var)

ggplot(dat, aes(index, total)) +
  geom_point() +
  geom_vline(xintercept = num, colour = "red") +
  labs(x = "PC", y = "Variance (%)") + 
  scale_x_continuous(n.breaks = 10, limits = c(0, NA)) +
  scale_y_continuous(n.breaks = 10, limits = c(0, NA)) +
  theme_minimal()
```

### Technical

Retain all PCs until the percentage of total variation explained reaches a given threshold based on the technical noise:

```{r}
set.seed(1939381666)

dec <- metadata(sce)$modelGeneVarWithSpikes

hvg <- rowData(sce)$selected

dim <- reducedDim(denoisePCA(sce, technical = dec, subset.row = hvg, BSPARAM = ExactParam()))

var <- attr(dim, "percentVar")

num <- ncol(dim)
```

Plot the cumulative percentage of variance explained by successive PCs:

```{r}
dat <- data.frame(index = seq_along(var), total = cumsum(var))

ggplot(dat, aes(index, total)) +
  geom_point() +
  geom_vline(xintercept = num, colour = "red") +
  labs(x = "PC", y = "Cumulative variance (%)") +
  theme_minimal()
```

### Structure

Find the number of subpopulations in the data using clustering:

```{r}
dim <- reducedDim(sce, "PCA")

fit <- getClusteredPCs(dim, min.rank = 1, max.rank = ncol(dim))

num <- metadata(fit)$chosen
```

Plot the percentage of variance explained by successive PCs:

```{r}
dat <- as.data.frame(fit)

ggplot(dat, aes(n.pcs, n.clusters)) +
  geom_point() + 
  geom_abline(slope = 1, intercept = 1, colour = "red") + 
  geom_vline(xintercept = num, colour = "grey", linetype = "dashed") +
  labs(x = "PC", y = "Number of clusters") +
  theme_minimal()
```

### Selection

Choose number of PCs to retain from the above analyses:

```{r}
num <- 3

len <- seq_len(num)

dim <- reducedDim(sce, "PCA")

reducedDim(sce, "PCA") <- dim[, len]
```

Plot scatterplot matrix of retained PCs:

```{r, fig.height = 8, fig.width = 10}
plotPCA(sce, colour_by = "cluster", ncomponents = num) + theme(legend.position = "none")
```

Cluster cells to help examine effect of TSNE and UMAP hyper-parameters:

```{r}
snn <- buildSNNGraph(sce, use.dimred = "PCA")

com <- igraph::cluster_walktrap(snn)

sce$cluster <- factor(com$membership)
```

## TSNE

Perform t-stochastic neighbour embedding (t-SNE) on the PCA data.

### Parameter exploration

Inspect TSNE plots generated by different perplexity and iteration parameters:

```{r, fig.height = 10, fig.width = 10}
lim <- (ncol(sce) - 1) / 3

max <- ifelse(lim < 50, lim, 50)

per <- seq(5, max, length.out = 5)

itr <- seq(500, 1500, length.out = 5)

arg <- expand.grid(
  perplexity = floor(per),
  max_iter = floor(itr)
)

plt <- apply(arg, 1, function(x) {

  set.seed(1667216890)

  run <- runTSNE(sce, dimred = "PCA", perplexity = x[1], max_iter = x[2])

  txt <- paste(x[1], x[2], sep = "/")

  plotTSNE(run, colour_by = "cluster") + ggtitle(txt) + 
    theme(axis.title = element_blank(), legend.position = "none")

})

wrap_plots(plt, nrow = 5, ncol = 5, byrow = FALSE)
```

### Calculate TSNE

Select optimal perplexity value and number of iterations:

```{r}
set.seed(1667216890)

sce <- runTSNE(sce, dimred = "PCA", perplexity = 14, max_iter = 1000)
```

## UMAP

Perform uniform manifold approximation and projection (UMAP) on the PCA data.

### Parameter exploration

Inspect UMAP plots generated by different neighbors and distance parameters:

```{r, fig.height = 10, fig.width = 10}
max <- ifelse(ncol(sce) < 100, ncol(sce), 100)

num <- seq(2, max, length.out = 5)

dst <- seq(0, 1, length.out = 5)

arg <- expand.grid(
  n_neighbors = floor(num),
  min_dist = signif(dst, digits = 2)
)

plt <- apply(arg, 1, function(x) {

  set.seed(1915759776)

  run <- runUMAP(sce, dimred = "PCA", n_neighbors = x[1], min_dist = x[2])

  txt <- paste(x[1], x[2], sep = "/")

  plotUMAP(run, colour_by = "cluster") + ggtitle(txt) + 
    theme(axis.title = element_blank(), legend.position = "none")

})

wrap_plots(plt, nrow = 5, ncol = 5, byrow = FALSE)
```

### Calculate UMAP

Select optimal number of nearest neighbors:

```{r}
set.seed(1915759776)

sce <- runUMAP(sce, dimred = "PCA", n_neighbors = 19, min_dist = 0.5)
```

## Variables {.tabset}

### PCA

```{r, fig.height = 8, fig.width = 10}
var <- c("sum", "detected", "subsets_MT_percent", "altexps_ERCC_percent")

plt <- lapply(var, function(x) plotPCA(sce, colour_by = x))

patchwork::wrap_plots(plt, ncol = 2)
```

### TSNE

```{r, fig.height = 8, fig.width = 10}
var <- c("sum", "detected", "subsets_MT_percent", "altexps_ERCC_percent")

plt <- lapply(var, function(x) plotTSNE(sce, colour_by = x))

patchwork::wrap_plots(plt, ncol = 2)
```

### UMAP

```{r, fig.height = 8, fig.width = 10}
var <- c("sum", "detected", "subsets_MT_percent", "altexps_ERCC_percent")

plt <- lapply(var, function(x) plotUMAP(sce, colour_by = x))

patchwork::wrap_plots(plt, ncol = 2)
```

## Summary

### Output

Write experiment data:

```{r output}
saveRDS(sce, file = "data/04-reduced-dimensions.rds")
```
